SUPER MESSY FIX
I have to make this nicer. The messy fix was to fix the problem where both ActionMaps were active at the same time so in order to fix it I have to only link up the other actionMap when it switches to that one. ONCE they both are linked and it has switched once it finally decides that only one actionMap can be active at a time. PLEASE GOD FIGURE THIS OUT SO I DON'T HAVE TO DO THIS!!!